Extract common record cursor code
